fix(docs): split Palette into separate components
and fixes for Chrome
